calories burned vs intake
The “caloriesBurnedVsIntake” argument in get_daily_data only returns
the burned calories. (Maybe fitbit changed something?). I changed the
function to return both caloriesBurned and caloriesIntake. I am not
sure if this is how you want to do it or if you want to split the
dataframe into two seperate dataframes.